+++ ChangeLog.rst
@@ -6,6 +6,7 @@ Features
 
 - Detect annotation and other options from the existing output files
 - Merge configuration from setup.cfg with requirements*.in files
+- Add "prequ check --verbose" for showing what is outdated
 
 Bug Fixes
 ~~~~~~~~~